feat: working on rbac

+++ app/Http/Controllers/Units/RanksController.php
@@ -11,6 +11,7 @@ use App\Http\Requests\Units\Ranks\CreateRankRequest;
 use App\Models\Rank;
 use App\Models\Unit;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Validation\ValidationException;
 use Inertia\Inertia;
 
@@ -18,6 +19,8 @@ class RanksController extends Controller
 {
     public function list(Unit $unit)
     {
+        Gate::authorize('viewAny', Rank::class);
+
         /** @var list<Rank> $ranks */
         $ranks = $unit
             ->ranks()
@@ -32,6 +35,8 @@ class RanksController extends Controller
 
     public function create(Unit $unit)
     {
+        Gate::authorize('create', Rank::class);
+
         $nextOrd = $unit->ranks()->max('ord') + 1;
 
         return Inertia::render('units/ranks/create', ['nextOrd' => $nextOrd]);
@@ -39,6 +44,8 @@ class RanksController extends Controller
 
     public function store(Unit $unit, CreateRankRequest $request, CreateNewRank $action)
     {
+        Gate::authorize('create', Rank::class);
+
         $action->create($unit, $request->post());
 
         Inertia::flash('toast', ['type' => 'success', 'message' => __('Successfully created rank.')]);
@@ -48,6 +55,8 @@ class RanksController extends Controller
 
     public function edit(Unit $unit, Rank $rank)
     {
+        Gate::authorize('update', $rank);
+
         return Inertia::render('units/ranks/edit', [
             'rank' => $rank,
         ]);
@@ -55,6 +64,8 @@ class RanksController extends Controller
 
     public function update(Unit $unit, Rank $rank, Request $request, UpdateRank $action)
     {
+        Gate::authorize('update', $rank);
+
         try {
             $action->update($rank, $request->post());
         } catch (\Exception $e) {
@@ -74,6 +85,8 @@ class RanksController extends Controller
 
     public function destroy(Unit $unit, Rank $rank, DeleteRank $action)
     {
+        Gate::authorize('destroy', $rank);
+
         try {
             $action->delete($rank);
         } catch (\Exception $e) {
